+++ src/test/java/seedu/address/logic/commands/DeleteProfileCommandTest.java
@@ -23,7 +23,7 @@ import seedu.address.model.profile.Profile;
  * Contains integration tests (interaction with the Model) and unit tests for
  * {@code DeleteCommand}.
  */
-public class DeleteCommandTest {
+public class DeleteProfileCommandTest {
 
     private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
